+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/Autos/AutoBlueA.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.Autos;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
@@ -9,25 +9,28 @@ import org.firstinspires.ftc.teamcode.toolkit.ULLinearOpMode;
 
 import java.util.ArrayList;
 
-@Autonomous(name = "RedBauto", group = "OpModes")
+@Autonomous(name = "AutoBlueA", group = "OpModes")
 
-public class RedBauto extends ULLinearOpMode {
+public class AutoBlueA extends ULLinearOpMode {
 
     Robot robot;
     Odometry odom;
 
     @Override
-    public void runOpMode(){
-        robot = new Robot();
-        odom = new Odometry(robot);
+    public void runOpMode() {
+        robot = new Robot ();
+        odom = new Odometry (robot);
 
         waitForStart();
 
         ArrayList<PathPoint> allPoints = new ArrayList<>();
 
-        allPoints.add(new PathPoint(0, 86, 0.7, 2, 5));
-        //allPoints.add(new PathPoint(20, 86, 0.3, 2, 5));
+        allPoints.add(new PathPoint(0, 72, 0.7, 2, 5));
+        //allPoints.add(new PathPoint(-38, 72, 0.5, 2, 5));
 
         odom.followPath(allPoints);
     }
 }
+
+
+
