+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/Autos/AutoRedA.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.Autos;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
@@ -9,27 +9,25 @@ import org.firstinspires.ftc.teamcode.toolkit.ULLinearOpMode;
 
 import java.util.ArrayList;
 
-@Autonomous(name = "RedCauto", group = "OpModes")
+@Autonomous(name = "AutoRedA", group = "OpModes")
 
-public class RedCauto extends ULLinearOpMode {
+public class AutoRedA extends ULLinearOpMode {
 
     Robot robot;
     Odometry odom;
 
     @Override
-    public void runOpMode(){
-        robot = new Robot();
-        odom = new Odometry(robot);
+    public void runOpMode() {
+        robot = new Robot ();
+        odom = new Odometry (robot);
 
         waitForStart();
 
         ArrayList<PathPoint> allPoints = new ArrayList<>();
 
-        allPoints.add(new PathPoint(0, 124, 0.7, 2, 5));
-        //allPoints.add(new PathPoint(42, 124, 0.5, 2, 5));
+        allPoints.add(new PathPoint(0, 72, 0.7, 2, 5));
+        //allPoints.add(new PathPoint(38, 72, 0.5, 2, 5));
 
         odom.followPath(allPoints);
-
     }
-
 }
