+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pathfiles/blueF.txt
@@ -1,0 +1,34 @@
+58 133
+58 126
+58 119
+58 112
+56 104
+51 99
+45 95
+39 92
+32 89
+25 86
+19 83
+14 77
+11 70
+9 64
+7 56
+7 49
+8 42
+12 35
+15 29
+20 22
+25 17
+31 13
+38 11
+45 14
+50 19
+55 25
+58 31
+62 37
+65 43
+67 50
+70 57
+73 64
+77 70
+82 75
