+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pathfiles/blueG.txt
@@ -1,0 +1,15 @@
+56 125
+55 118
+55 111
+54 104
+54 97
+53 90
+53 82
+52 75
+51 68
+49 61
+42 60
+35 59
+28 58
+20 56
+13 55
