+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pathfiles/blueI.txt
@@ -1,0 +1,15 @@
+55 133
+55 126
+55 119
+55 111
+55 104
+55 96
+55 89
+55 82
+54 75
+54 68
+50 62
+43 61
+35 61
+28 61
+21 61
